(#116) Move Assertion2 -> Assertion
