+++ src/test/java/org/llorllale/cactoos/matchers/HasSizeTest.java
@@ -48,7 +48,7 @@ public final class HasSizeTest {
 
     @Test
     public void matchesIterableSize() {
-        new Assertion2<>(
+        new Assertion<>(
             "matches iterable with given size",
             new HasSize(2),
             new Matches<>(new IterableOfBooleans(true, false))
@@ -57,7 +57,7 @@ public final class HasSizeTest {
 
     @Test
     public void doesNotMatchIterableSize() {
-        new Assertion2<>(
+        new Assertion<>(
             "does not match an iterable with a different size",
             new IsNot<>(new HasSize(2)),
             new Matches<>(new IterableOfInts(1))
@@ -66,7 +66,7 @@ public final class HasSizeTest {
 
     @Test
     public void matchesEmptyCollection() {
-        new Assertion2<>(
+        new Assertion<>(
             "matches empty iterable if given size arg is 0",
             new HasSize(0),
             new Matches<>(new ListOf<>())
@@ -77,7 +77,7 @@ public final class HasSizeTest {
     public void describesMismatch() {
         final Description description = new StringDescription();
         new HasSize(2).describeMismatchSafely(new ListOf<>(), description);
-        new Assertion2<>(
+        new Assertion<>(
             "describes mismatch",
             new TextOf(description.toString()),
             new TextIs(
